Update Excel2Word.bas
Removed some references that would have required to add word as a dependency to the Macro.

Removed hard coded report title. Rather it will use a named range called `report_title`.

Moved initialization of Word application up, so the code does not fail if Word is not open when the OLE Object is activated.